# docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "API for Boston Archery",
    "title": "Boston Archery API",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "paths": {
    "/player": {
      "post": {
        "tags": [
          "player"
        ],
        "summary": "Add a new player.",
        "operationId": "postPlayer",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "firstName",
                "lastName"
              ],
              "properties": {
                "firstName": {
                  "type": "string",
                  "x-go-name": "FirstName"
                },
                "lastName": {
                  "type": "string",
                  "x-go-name": "LastName"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PostPlayerResp",
            "schema": {
              "$ref": "#/definitions/PostPlayerResp"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/player/{ID}": {
      "get": {
        "tags": [
          "player"
        ],
        "summary": "Get a player by ID.",
        "operationId": "getPlayer",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "GetPlayerResp",
            "schema": {
              "$ref": "#/definitions/GetPlayerResp"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": [
          "player"
        ],
        "summary": "Update a player.",
        "operationId": "putPlayer",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "firstName",
                "lastName"
              ],
              "properties": {
                "firstName": {
                  "type": "string",
                  "x-go-name": "FirstName"
                },
                "lastName": {
                  "type": "string",
                  "x-go-name": "LastName"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PutPlayerResp",
            "schema": {
              "$ref": "#/definitions/PutPlayerResp"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/players": {
      "get": {
        "tags": [
          "player"
        ],
        "summary": "Get all players.",
        "operationId": "getPlayers",
        "responses": {
          "200": {
            "description": "GetPlayersResp",
            "schema": {
              "$ref": "#/definitions/GetPlayersResp"
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/season": {
      "post": {
        "tags": [
          "season"
        ],
        "summary": "Add a new season.",
        "operationId": "postSeason",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "startDate",
                "endDate"
              ],
              "properties": {
                "byeWeeks": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "date"
                  },
                  "x-go-name": "ByeWeeks"
                },
                "endDate": {
                  "type": "string",
                  "format": "date",
                  "x-go-name": "EndDate"
                },
                "name": {
                  "type": "string",
                  "x-go-name": "Name"
                },
                "startDate": {
                  "type": "string",
                  "format": "date",
                  "x-go-name": "StartDate"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PostSeasonResp",
            "schema": {
              "$ref": "#/definitions/PostSeasonResp"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/season/{ID}": {
      "put": {
        "tags": [
          "season"
        ],
        "summary": "Update a season.",
        "operationId": "putSeason",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "startDate",
                "endDate"
              ],
              "properties": {
                "byeWeeks": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "date"
                  },
                  "x-go-name": "ByeWeeks"
                },
                "endDate": {
                  "type": "string",
                  "format": "date",
                  "x-go-name": "EndDate"
                },
                "name": {
                  "type": "string",
                  "x-go-name": "Name"
                },
                "startDate": {
                  "type": "string",
                  "format": "date",
                  "x-go-name": "StartDate"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PutSeasonResp",
            "schema": {
              "$ref": "#/definitions/PutSeasonResp"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/seasons": {
      "get": {
        "tags": [
          "season"
        ],
        "summary": "Get all seasons.",
        "operationId": "getSeasons",
        "parameters": [
          {
            "type": "string",
            "name": "Name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "GetSeasonsResp",
            "schema": {
              "$ref": "#/definitions/GetSeasonsResp"
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "type": "object",
      "properties": {
        "Msg": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/errors"
    },
    "GetPlayerResp": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Player"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/players"
    },
    "GetPlayersResp": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Player"
          },
          "x-go-name": "Data"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/players"
    },
    "GetSeasonsResp": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Season"
          },
          "x-go-name": "Data"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/seasons"
    },
    "Player": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "x-go-name": "FirstName"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "lastName": {
          "type": "string",
          "x-go-name": "LastName"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/players"
    },
    "PostPlayerResp": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Player"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/players"
    },
    "PostSeasonResp": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Season"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/seasons"
    },
    "PutPlayerResp": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Player"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/players"
    },
    "PutSeasonResp": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Season"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/seasons"
    },
    "Season": {
      "type": "object",
      "properties": {
        "byeWeeks": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "ByeWeeks"
        },
        "endDate": {
          "type": "string",
          "x-go-name": "EndDate"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "startDate": {
          "type": "string",
          "x-go-name": "StartDate"
        }
      },
      "x-go-package": "github.com/mellena1/boston-archery-api/handlers/seasons"
    }
  }
}